backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
asyncpg==0.29.0
pyjwt==2.8.0
python-telegram-bot==20.8
pydantic==2.6.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.26.0
